requirements.txt
langchain==0.2.10
langchain-community==0.2.10
langchain-core==0.2.10

faiss-cpu
sentence-transformers
pypdf
python-dotenv
streamlit
groq
